fix: mock 用の処理追加
Refs: #21

+++ src/frontend/apps/vuejs/myapp/myapp/src/components/resource01/DeleteBook.vue
@@ -20,22 +20,27 @@ export default {
 
             const url = this.baseUrl + '/' + this.id
 
-            fetch(url, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': 'Bearer ' + this.token,
-                }
-            }).then(response => {
-                console.log('Success:', response)
+            if (import.meta.env.VITE_MOCK_MODE) {
+                this.result['status'] = 204
+                this.result['statusText'] = 'dummy response.statusText'
+            } else {
+                fetch(url, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Authorization': 'Bearer ' + this.token,
+                    }
+                }).then(response => {
+                    console.log('Success:', response)
 
-                this.result['status'] = response.status
-                this.result['statusText'] = response.statusText
-            }).catch(error => {
-                console.error('Error:', error)
+                    this.result['status'] = response.status
+                    this.result['statusText'] = response.statusText
+                }).catch(error => {
+                    console.error('Error:', error)
 
-                this.result['error'] = error
-            })
+                    this.result['error'] = error
+                })
+            }
         }
     }
 }
